Implement centralized network caching and add caching for space events and NASA API.
- Created `apts/utils/network.py` with `get_session` using `requests_cache`.
- Refactored weather providers and light pollution to use the new central session.
- Added caching to space launches, space events, and NASA comet data to prevent rate-limiting errors.
- Updated unit tests for space events to reflect the new caching implementation.
